Create Grades+ID.cs
Write a program that reads in the Student ID (integer) and scores (double, out of 100) of 30 students and stores them in two arrays. Then the programs adds 2 points to all students whose score are less than 50. Next, the program displays the Student ID with the highest grade as well the percentage of students who passed the course