Create Shop_array
Write a program that creates 4 arrays for the id, price, SN and quantity of 1000 items being sold in a shop. The program prints the data in a table format. The program then prints out the ids of all items whose quantity is less than 10, and those items whose price is greater than 500.