Create tip.cs
Write a program to calculate the tip a waiter should get in a
restaurant. The input to the program is the number of people in the group and the bill. The tip is calculated as follows: if the bill is less than 50 pounds then the tip is fixed at 2 pounds per person in the group. If the bill is between 50 and 300 pounds then the tip is 10% of the bill plus 3 pound per person. If the bill is above 300 pounds then the tip is 15% of the bill. The output of the program should be the tip AND the total amount to be paid.